Move records package to legacy.
