+++ politeiawww/legacy/politeiawww.go
@@ -34,13 +34,13 @@ import (
 	ghtracker "github.com/decred/politeia/politeiawww/legacy/codetracker/github"
 	"github.com/decred/politeia/politeiawww/legacy/comments"
 	"github.com/decred/politeia/politeiawww/legacy/pi"
+	"github.com/decred/politeia/politeiawww/legacy/records"
 	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
 	"github.com/decred/politeia/politeiawww/legacy/user"
 	"github.com/decred/politeia/politeiawww/legacy/user/cockroachdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/localdb"
 	"github.com/decred/politeia/politeiawww/legacy/user/mysql"
 	"github.com/decred/politeia/politeiawww/mail"
-	"github.com/decred/politeia/politeiawww/records"
 	"github.com/decred/politeia/politeiawww/sessions"
 	"github.com/decred/politeia/politeiawww/websockets"
 	"github.com/decred/politeia/util"
