+++ politeiawww/legacy/routes.go
@@ -16,8 +16,8 @@ import (
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/politeiawww/legacy/comments"
 	"github.com/decred/politeia/politeiawww/legacy/pi"
+	"github.com/decred/politeia/politeiawww/legacy/records"
 	"github.com/decred/politeia/politeiawww/legacy/ticketvote"
-	"github.com/decred/politeia/politeiawww/records"
 	"github.com/decred/politeia/util"
 )
